==> code/2-exploration.Rmd <==
---
title: "2-exploration"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message = FALSE}
# load packages
library(tidyverse)  # tidyverse
library(ggrepel)    # for scatter plot point labels
library(kableExtra) # for printing tables
library(cowplot)    # for side by side plots
library(corrplot) # correlation plots
library(dplyr)
library(readr)
library(stringr)
```

```{r}
# load cleaned data
pdiet_data = read.csv("./../data/clean/pregnancy_data.csv")
```

```{r}
# calculate median weight gain
median_weight_gain = pdiet_data %>%
  summarise(median(pregnancy_weight_gain)) %>%
  pull()

# create histogram of weight gain during pregnancy
hist = pdiet_data %>%
  ggplot(aes(x = pregnancy_weight_gain)) + 
  geom_histogram() +
  geom_vline(xintercept = median_weight_gain,
             linetype = "dashed") +
  labs(x = "Weight gain during pregnancy", 
       y = "Number of women") +
  theme_classic()

# save
ggsave(filename = "~/pregnancy-weight-gain-final-project/results/weight-gain-hist.png", 
       plot = hist, 
       device = "png", 
       width = 7, 
       height = 4)

```


```{r}
# scatter plots of weight gain vs age, bmi, poverty/income ratio
age_scatter = pdiet_data %>% ggplot(aes(x=mother_age, y=pregnancy_weight_gain)) + geom_point() +
  labs(x="age", y="pregnancy weight gain") + theme_bw()

bmi_scatter = pdiet_data %>% ggplot(aes(x=prepregnancy_bmi, y=pregnancy_weight_gain)) + geom_point() +
  labs(x="BMI before pregnancy", y="pregnancy weight gain") + theme_bw()

povincome_scatter = pdiet_data %>% ggplot(aes(x=poverty_income_ratio, y=pregnancy_weight_gain)) + geom_point() +
  labs(x="poverty-income ratio", y="pregnancy weight gain") + theme_bw()

# save
ggsave(filename = "~/pregnancy-weight-gain-final-project/results/age-scatter.png", 
       plot = age_scatter, 
       device = "png", 
       width = 7, 
       height = 4)

ggsave(filename = "~/pregnancy-weight-gain-final-project/results/bmi-scatter.png", 
       plot = bmi_scatter, 
       device = "png", 
       width = 7, 
       height = 4)

ggsave(filename = "~/pregnancy-weight-gain-final-project/results/povincome-scatter.png", 
       plot = povincome_scatter, 
       device = "png", 
       width = 7, 
       height = 4)

scatters = plot_grid(age_scatter, bmi_scatter, povincome_scatter)

ggsave(filename = "~/pregnancy-weight-gain-final-project/results/scatters.png", 
       plot = scatters, 
       device = "png", 
       width = 7, 
       height = 4)
```

```{r}
# correlation matrices for DHP variables
# subset only DHP variables
pdiet_DHP = pdiet_data[,substr(colnames(pdiet_data), 1, 3)=="DHP"]

# calculate correlation and make plot
corr = cor(pdiet_DHP)

png(height=1800, width=1800, file="~/pregnancy-weight-gain-final-project/results/DHP-correlation.png", type = "cairo")
corrplot(corr, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45, tl.cex = 0.7)
dev.off()

flattenCorrMatrix <- function(cormat) {
  ut <- upper.tri(cormat)
  data.frame(
    row = rownames(cormat)[row(cormat)[ut]],
    column = rownames(cormat)[col(cormat)[ut]],
    cor  =(cormat)[ut]
    )
}
# top 10 correlated DHP variables
top10 = flattenCorrMatrix(corr) %>% slice_max(cor, n=10)
```
