==> code/3-train-test-split.Rmd <==
---
title: "3-train-test-split"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readr)
```

```{r}
# load cleaned data
pdiet_data = read.csv("./../data/clean/pregnancy_data.csv")

# split into train and test
set.seed(5) # seed set for reproducibility (DO NOT CHANGE)
n = nrow(pdiet_data)
train_samples = sample(1:n, round(0.8*n))
# split data into training and test sets
pdiet_train = pdiet_data[train_samples,]
pdiet_test = pdiet_data[-train_samples,]

# save the train and test data
write_csv(x = pdiet_train, file = "./../data/clean/pdiet_train.csv")
write_csv(x = pdiet_test, file = "./../data/clean/pdiet_test.csv")
```
