==> code/4-regression-modeling.Rmd <==
---
title: "4-regression-modeling"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Regression Modeling 
```{r}
#load libraries
library(glmnetUtils)  #for ridge regression
library(readr)
source("functions/plot_glmnet.R")   #for ridge trace plot

```

```{r}
# load in training data
pdiet_train = read_csv("./../data/clean/pdiet_train.csv", show_col_types = FALSE)
```

```{r, warning=FALSE}
# run linear regression on DHP variables (numeric)
lm_fit = lm(formula = pregnancy_weight_gain ~ ., data = pdiet_train)

# print summary
summary(lm_fit)
res = resid(lm_fit) %>% as.numeric()
typeof(res)

# residual histogram
res_hist = ggplot(data = pdiet_train, aes(x = lm_fit$residuals)) +
    geom_histogram() + theme_bw() + labs(x= "residuals")

# save plot
ggsave(filename = "./../results/residuals-hist.png", 
       plot = res_hist, 
       device = "png", 
       width = 6, 
       height = 4)

# save linear regression fit
save(lm_fit, file = "./../results/lm_fit.Rda")
```



```{r}
# run ridge regression
set.seed(5)
ridge_fit = cv.glmnet( pregnancy_weight_gain ~ .,
alpha = 0,
nfolds = 10,
data = pdiet_train)

#save ridge fit
save(ridge_fit, file = "./../results/ridge_fit.Rda")

# plot ridge fit (CV plot)
plot(ridge_fit)

# save ridge CV plot
png(width = 6, 
    height = 4,
    res = 300,
    units = "in", 
    filename = "./../results/ridge-cv-plot.png")
plot(ridge_fit)
dev.off()

#plot trace plot 
l_ridge_fit = plot_glmnet(ridge_fit, pdiet_train, features_to_plot = 6)

# save ridge trace plot
ggsave(filename = "./../results/ridge-trace-plot.png", 
       plot = l_ridge_fit, 
       device = "png", 
       width = 6, 
       height = 4)
```

```{r}
# evaluate training RMSEs

# evaluate OLSR RMSE
lm_predictions = predict(lm_fit, newdata = pdiet_train)

lm_RMSE = sqrt(mean((lm_predictions-pdiet_train$pregnancy_weight_gain)^2))

# evaluate ridge RMSE
ridge_predictions = predict(ridge_fit, 
                            newdata = pdiet_train, 
                            s = "lambda.1se") %>%
  as.numeric()
ridge_RMSE = sqrt(mean((ridge_predictions-pdiet_train$pregnancy_weight_gain)^2))

```
