==> code/5-tree-modeling.Rmd <==
---
title: "5-tree-modeling"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message = FALSE}
# load packages
library(rpart)         # to train decision trees
library(rpart.plot)    # to plot decision trees
library(randomForest)  # random forests
library(gbm)           # boosting
library(tidyverse)     # tidyverse
library(kableExtra)
library(ggplot2) 
library(readr)
```

```{r}
# load in training and test data
pregnancy_train = read.csv("./../data/clean/pdiet_train.csv")
pregnancy_test = read.csv("./../data/clean/pdiet_test.csv")
```

```{r}
# run random forest with default parameters
set.seed(5) # for reproducibility (DO NOT CHANGE)
rf_fit = randomForest(pregnancy_weight_gain ~ ., data = pregnancy_train)
error_B = plot(rf_fit)
plot(rf_fit)
# save
png(file='~/pregnancy-weight-gain-final-project/results/error-B.png')
plot(rf_fit)
dev.off() 

# OOB error stabilizes at approx 100 trees based on plot
# default mtry value: p/3, or 31/3, which is 11

# tune for optimal value of m
mvalues = seq(1,31)
oob_errors = numeric(length(mvalues))
ntree = 100
for(idx in 1:length(mvalues)){
  m = mvalues[idx]
  rf_fit = randomForest(pregnancy_weight_gain ~ ., mtry = m, data = pregnancy_train)
  oob_errors[idx] = rf_fit$mse[ntree]
}

# make plot of error vs m
error_m = tibble(m = mvalues, oob_err = oob_errors) %>%
  ggplot(aes(x = m, y = oob_err)) + 
  geom_line() + geom_point() + 
  scale_x_continuous(breaks = mvalues, guide = guide_axis(check.overlap = TRUE)) +
  theme_bw()

# observe m minimized at 28

# save
ggsave(filename = "~/pregnancy-weight-gain-final-project/results/error-versus-m.png", 
       plot = error_m, 
       device = "png", 
       width = 7, 
       height = 4)
```

```{r}
# fit model with optimal m and B
set.seed(5) # for reproducibility (DO NOT CHANGE)

rfit_final = randomForest(pregnancy_weight_gain ~ ., mtry = 28, ntree = 100, importance = TRUE, data = pregnancy_train)

#save random forest fit
save(rfit_final, file = "./../results/rf_fit.Rda")

# variable importance plots
png(file='~/pregnancy-weight-gain-final-project/results/var-imp.png')
varImpPlot(rfit_final, n.var = 15)
dev.off() 

```
```{r}
# calculate train rmse
train_rf_probabilities = predict(rfit_final, newdata = pregnancy_train)
train_RMSE_rf = sqrt(mean((train_rf_probabilities - pregnancy_train$pregnancy_weight_gain)^2))
```
