==> code/6-model-evaluation.Rmd <==
---
title: "6-model-evaluation"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Model Evaluation

```{r}
# load libraries
library(glmnetUtils)
library(tidyverse)
library(kableExtra)
library(readr)
```

```{r}
# load test data 

pdiet_test = read_csv("./../data/clean/pdiet_test.csv", show_col_types = FALSE)

```

```{r}
#load linear regression fit
load("./../results/lm_fit.Rda")

# load ridge fit
load("./../results/ridge_fit.Rda")

# load random forest fit
load("./../results/rf_fit.Rda")


```

```{r}
# evaluate linear regression RMSE
linear_predictions = predict(lm_fit, newdata = pdiet_test)
linear_RMSE = sqrt(mean((linear_predictions - pdiet_test$pregnancy_weight_gain)^2))

# evaluate ridge regression RMSE
ridge_predictions = predict(ridge_fit, 
                            newdata = pdiet_test, 
                            s = "lambda.1se") %>% as.numeric()
ridge_RMSE = sqrt(mean((ridge_predictions - pdiet_test$pregnancy_weight_gain)^2))

# evaluate random forest RMSE
rf_predictions = predict(rfit_final, newdata = pdiet_test)
rf_RMSE = sqrt(mean((rf_predictions - pdiet_test$pregnancy_weight_gain)^2))

#make table
error_for_models = tribble(
  ~Model, ~RMSE, 
  #------/------- 
  "Linear", linear_RMSE,
  "Ridge", ridge_RMSE,
  "Random Forest", rf_RMSE
  )
error_for_models
# make nice table
error_for_models %>% kable(format = "latex", row.names = NA, 
                        booktabs = TRUE,
                        digits = 5,
                        col.names = c("Model type", 
                                      "Root mean squared error"),
                        caption = "These are the root mean squared test 
                        errors of the linear model, ridge regression, and random forest") %>%
  kable_styling(position = "center") %>%
  kable_styling(latex_options = "HOLD_position")

tibble(Method = c("OLSR", "Ridge", "Random Forest"), `Test RMSE` = c(linear_RMSE, ridge_RMSE, rf_RMSE)) %>%
  write_tsv("./../results/model-evaluation.tsv")

```
